[Refactor] next font 적용
