+++ src/api/Fonts/GugiFont.js
@@ -1,7 +1,7 @@
-import localFont from "@/next/font/local";
+import localFont from 'next/font/local';
 
 const gugiFont = localFont({
-  src: './fonts/Gugi/Gugi-Regular.ttf',
-  display: 'swap'
-})
+  src: './../../app/fonts/Gugi/Gugi-Regular.ttf',
+  display: 'swap',
+});
 export default gugiFont;
